+++ tests/manual/create-hlr.ts
@@ -1,3 +1,7 @@
 import { client, print, value } from './_setup';
 
-print(await client().createHlr(value('EDA_MSISDN', 2), value('EDA_IMSI', 3)));
+print(
+  await client().createHlr(value('EDA_MSISDN', 2), value('EDA_IMSI', 3), {
+    paidMode: 'PREPAID',
+  }),
+);